requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
deepgram-sdk>=3.0.0,<4.0.0
anthropic>=0.75.0
openai>=2.14.0
python-multipart==0.0.9
python-dotenv==1.0.1
numpy==1.26.4
aiofiles==24.1.0
pypdf==6.5.0
python-docx==1.2.0
openpyxl==3.1.5
pillow>=10.0.0
markdown>=3.4
PyJWT>=2.8.0
cryptography>=41.0.0
httpx>=0.25.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
pgvector>=0.2.4
redis>=5.0.0
pinecone>=3.0.0
